Adds Sitemap concern
